# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: string;
  name: string;
  icon: string;
  count: number;
}

export interface Sucursal {
  id: string;
  name: string;
  address: string;
  city: string;
  hours: string;
}

export interface Spec {
  label: string;
  value: string;
}

export interface Product {
  id: string;
  sku: string;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  stock: number;
  /** @nullable */
  categoryId?: string | null;
  /** @nullable */
  image?: string | null;
  compatible: boolean;
  specs: Spec[];
  vehicles: string[];
  /** @nullable */
  oem?: string[] | null;
  /** @nullable */
  equivalents?: string[] | null;
}

export interface ProductPage {
  items: Product[];
  total: number;
}

export interface Deals {
  dealOfDay?: Product | null;
  ofertas: Product[];
}

export interface SyncStatus {
  status: string;
  /** @nullable */
  lastStartedAt?: string | null;
  /** @nullable */
  lastFinishedAt?: string | null;
  /** @nullable */
  lastSuccessAt?: string | null;
  /** @nullable */
  lastError?: string | null;
  /** @nullable */
  message?: string | null;
  productsSynced: number;
  categoriesSynced: number;
  sucursalesSynced: number;
  brandsSynced: number;
}

export interface OrderLineInput {
  productId: string;
  sku: string;
  name: string;
  qty: number;
  price: number;
}

export interface OrderInput {
  lines: OrderLineInput[];
  sucursalId: string;
  entrega: string;
  pago: string;
  total: number;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  buyerPhone?: string | null;
}

export interface OrderResult {
  id: number;
  folio: string;
  status: string;
  /** @nullable */
  admintotalPedidoId?: string | null;
}

export type ListProductsParams = {
q?: string;
categoryId?: string;
brand?: string;
/**
 * When set, stock is reported for this sucursal only.
 */
sucursalId?: string;
limit?: number;
offset?: number;
};

export type GetProductParams = {
sucursalId?: string;
};

export type GetDealsParams = {
sucursalId?: string;
};