# lib/api-zod/src/generated/types/deals.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface Deals {
  dealOfDay?: Product | null;
  ofertas: Product[];
}