# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineInput } from './orderLineInput';

export interface OrderInput {
  lines: OrderLineInput[];
  sucursalId: string;
  entrega: string;
  pago: string;
  total: number;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  buyerPhone?: string | null;
}