# lib/api-zod/src/generated/types/orderResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderResult {
  id: number;
  folio: string;
  status: string;
  /** @nullable */
  admintotalPedidoId?: string | null;
}