# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Spec } from './spec';

export interface Product {
  id: string;
  sku: string;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  stock: number;
  /** @nullable */
  categoryId?: string | null;
  /** @nullable */
  image?: string | null;
  compatible: boolean;
  specs: Spec[];
  vehicles: string[];
  /** @nullable */
  oem?: string[] | null;
  /** @nullable */
  equivalents?: string[] | null;
}